backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==11.0.0
plaid-python==14.0.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
pandas==2.2.3
python-dateutil==2.9.0
openpyxl==3.1.5
